+++ CHANGELOG.md
@@ -10,6 +10,7 @@ All notable changes to this project will be documented in this file.
 - [#2319] AS3 direct editation - class gets removed after pressing cancel
 - [#2320] AS3 direct editation - modified flag of scripts vanishes after editing other script with and having error
 - [#2272] Filters strength attribute caps at 100%
+- [#2322] AS3 Construct property name formatting
 
 ## [21.1.0] - 2024-09-23
 ### Added
@@ -3582,6 +3583,7 @@ Major version of SWF to XML export changed to 2.
 [#2319]: https://www.free-decompiler.com/flash/issues/2319
 [#2320]: https://www.free-decompiler.com/flash/issues/2320
 [#2272]: https://www.free-decompiler.com/flash/issues/2272
+[#2322]: https://www.free-decompiler.com/flash/issues/2322
 [#943]: https://www.free-decompiler.com/flash/issues/943
 [#1812]: https://www.free-decompiler.com/flash/issues/1812
 [#2287]: https://www.free-decompiler.com/flash/issues/2287